requirements_python39.txt
pytest==7.4.2
numpy==1.24.3
torch==2.0.1
tensorflow==2.10.0
tensorflow-hub==0.14.0
einops==0.6.1
opencv-python==4.8.0.76
transformers==4.33.2
torchvision==0.15.2
imageio==2.31.6
sympy==1.12
scikit-image==0.21.0
lmdb==1.4.1
tqdm==4.65.2
requests==2.31.0
scipy==1.11.4
matplotlib==3.7.5
pillow==10.0.1
torchsde==0.2.6
six==1.16.0
attrs==23.1.0
albumentations==1.3.1
imagecorruptions==1.1.2
wheel==0.42.0
pyaml==25.1.0
